backup - refactors runner to move to an 'Observer pattern' for the Analyser
Runner is an Ochestrator: if any structure changes, it changes, and it's impossible to change a behaviour without changing the runner or the step implementation.

'Observer Pattern' is adding an 'Observer' (or two) for each steps. Logging, error collection, interruption, channeling returns to next step, ... are all becoming implementations of an Observer of each steps.

Notes:
* skipping mis-analysed media doesn't skip the progress or the error collection (but it won't be interrupted)
* Runner is taking an Option as parameter, refactoring not completed yet (lot of parameters still inlined)
* only Analyser has been refactored to become "Observable"